Develop

+++ app/domain/models/location_model.py
@@ -1,9 +1,8 @@
 from uuid import UUID
-from decimal import Decimal
 
 
 class LocationModel:
-    def __init__(self, id: UUID | None, incident_id: UUID | None, latitude: Decimal, longitude: Decimal, direction: str):
+    def __init__(self, id: UUID | None, incident_id: UUID | None, latitude: float, longitude: float, direction: str):
         self.id = id
         self.incident_id = incident_id
         self.latitude = latitude
